requirements.txt
Flask==3.1.2
pymongo==4.16.0
python-dotenv==1.2.1
